Consistently use pointer receivers for core.Number
Change all occurrences of value to pointer receivers
Add meta sys files to .gitignore
Code cleanup e.g.
- Don't capitalize error statements
- Fix ignored errors
- Fix ambiguous variable naming
- Remove unnecessary type casting
- Use named params

Fix #306

+++ exporter/trace/jaeger/jaeger_test.go
@@ -155,7 +155,7 @@ func Test_spanDataToThrift(t *testing.T) {
 	messageEventValue := "event-test"
 	keyValue := "value"
 	statusCodeValue := int64(2)
-	doubleValue := float64(123.456)
+	doubleValue := 123.456
 	boolTrue := true
 	statusMessage := "Unknown"
